mmap: Add file descriptor as parameter to PATH
A programmer may wish to create a shared memory object via the POSIX shm
API, mmap a descriptor of an anonyous file through some local address
socket, or otherwise mmap a file descriptor whose pathname cannot or
should not be determined.

This commit enables file descriptors to be passed through the PATH
parameter of MMAP on Unix systems. Since it did not open the fd, MUNMAP
does not close the fd either.

Although fstat *could* be used to automatically determine the size of
the file descriptor, this pass goes for the simpler option and requires
the SIZE parameter.